pyproject.toml
[tool.poetry]
name = "uw-idp-web-tests"
version = "0.1.0"
description = ""
authors = ["Tom Thorogood <goodtom@uw.edu>", "Glen Goulter <glenrg@uw.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
black = "^21.5b2"
boto3 = "^1.17.92"
click = "^8.0.1"
google-cloud-secret-manager = "^2.5.0"
inflection = "^0.5.1"
nslookup = "^1.4.0"
pydantic = "^1.8.2"
pytest = "^6.2.4"
PyYAML = "^5.4.1"
typing-extensions = "^3.10.0"
uw-webdriver-recorder = ">=4.0.0,<4.1.0"
waiter = "^1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
